qr and urgency changes

+++ src/main/java/in/handyman/raven/lib/model/qrextraction/QrOutputEntity.java
@@ -18,7 +18,7 @@ import java.util.stream.Stream;
 public class QrOutputEntity implements CoproProcessor.Entity {
     private String originId;
     private Integer groupId;
-    private Integer paperNo;
+    private Long paperNo;
     private String fileId;
     private String qrFormat;
     private Integer qrFormatId;
